Incr 8 - Перенес общие пакеты в common
