+++ internal/agent/collector/collector.go
@@ -1,6 +1,8 @@
 package collector
 
-import "github.com/m1khal3v/gometheus/internal/metric"
+import (
+	"github.com/m1khal3v/gometheus/internal/common/metric"
+)
 
 type Collector interface {
 	Collect() []metric.Metric
