+++ internal/agent/collector/random/random.go
@@ -2,8 +2,8 @@ package random
 
 import (
 	"fmt"
-	"github.com/m1khal3v/gometheus/internal/metric"
-	"github.com/m1khal3v/gometheus/internal/metric/kind/gauge"
+	"github.com/m1khal3v/gometheus/internal/common/metric"
+	"github.com/m1khal3v/gometheus/internal/common/metric/kind/gauge"
 	"math/rand/v2"
 )
