+++ internal/agent/collector/runtime/runtime.go
@@ -2,10 +2,10 @@ package runtime
 
 import (
 	"fmt"
-	"github.com/m1khal3v/gometheus/internal/logger"
-	"github.com/m1khal3v/gometheus/internal/metric"
-	"github.com/m1khal3v/gometheus/internal/metric/kind/counter"
-	"github.com/m1khal3v/gometheus/internal/metric/kind/gauge"
+	"github.com/m1khal3v/gometheus/internal/common/logger"
+	"github.com/m1khal3v/gometheus/internal/common/metric"
+	"github.com/m1khal3v/gometheus/internal/common/metric/kind/counter"
+	"github.com/m1khal3v/gometheus/internal/common/metric/kind/gauge"
 	"reflect"
 	"runtime"
 )
