+++ internal/agent/storage/storage.go
@@ -1,7 +1,7 @@
 package storage
 
 import (
-	"github.com/m1khal3v/gometheus/internal/metric"
+	"github.com/m1khal3v/gometheus/internal/common/metric"
 	"sync"
 )
