+++ internal/server/config/config.go
@@ -6,14 +6,20 @@ import (
 )
 
 type Config struct {
-	Address  string `env:"ADDRESS"`
-	LogLevel string `env:"LOG_LEVEL"`
+	Address         string `env:"ADDRESS"`
+	LogLevel        string `env:"LOG_LEVEL"`
+	StoreInterval   uint32 `env:"STORE_INTERVAL"`
+	FileStoragePath string `env:"FILE_STORAGE_PATH"`
+	Restore         bool   `env:"RESTORE"`
 }
 
 func ParseConfig() *Config {
 	config := &Config{}
 	flag.StringVarP(&config.Address, "address", "a", "localhost:8080", "address of gometheus server")
 	flag.StringVarP(&config.LogLevel, "log-level", "l", "info", "log level")
+	flag.Uint32VarP(&config.StoreInterval, "store-interval", "i", 300, "dump metrics to file interval in seconds")
+	flag.StringVarP(&config.FileStoragePath, "file-storage-path", "f", "/tmp/metrics-db.json", "file storage path")
+	flag.BoolVarP(&config.Restore, "restore", "r", true, "restore metrics from file")
 	flag.Parse()
 	if err := env.Parse(&config); err != nil {
 		panic(err)
