+++ internal/server/server.go
@@ -1,7 +1,7 @@
 package server
 
 import (
-	"github.com/m1khal3v/gometheus/internal/logger"
+	"github.com/m1khal3v/gometheus/internal/common/logger"
 	"github.com/m1khal3v/gometheus/internal/server/router"
 	"github.com/m1khal3v/gometheus/internal/server/storage/memory"
 	"net/http"
