+++ internal/server/server_test.go
@@ -2,10 +2,10 @@ package server
 
 import (
 	"fmt"
-	"github.com/m1khal3v/gometheus/internal/metric"
-	"github.com/m1khal3v/gometheus/internal/metric/factory"
-	"github.com/m1khal3v/gometheus/internal/metric/kind/counter"
-	"github.com/m1khal3v/gometheus/internal/metric/kind/gauge"
+	"github.com/m1khal3v/gometheus/internal/common/metric"
+	"github.com/m1khal3v/gometheus/internal/common/metric/factory"
+	"github.com/m1khal3v/gometheus/internal/common/metric/kind/counter"
+	"github.com/m1khal3v/gometheus/internal/common/metric/kind/gauge"
 	"github.com/m1khal3v/gometheus/internal/server/router"
 	"github.com/m1khal3v/gometheus/internal/server/storage/memory"
 	"github.com/stretchr/testify/assert"
